requirements.txt
langchain
langgraph
langchain-ollama
python-dotenv
pydantic